+++ core/token.h
@@ -10,9 +10,9 @@ struct token {
     char* text;
 
     /* token type identifier constants, sequentially defined starting
-         from EOF and ending with the FAIL token */
+       from EOF and ending with the FAIL token */
     static const uint END     = 0;
-    static const uint INT     = 1;  // integer string sequence
+    static const uint INTEGER = 1;  // integer string sequence
     static const uint EQ      = 2;  // =
     static const uint OPEN    = 3;  // {
     static const uint CLOSE   = 4;  // }
@@ -21,7 +21,7 @@ struct token {
     static const uint DATE    = 7;  // date string of the form 867.1.1, loosely
     static const uint QDATE   = 8;  // quoted date
     static const uint COMMENT = 9;  // hash-style comment (starts with '#')
-    static const uint FLOAT   = 10;
+    static const uint DECIMAL = 10;
     static const uint FAIL    = 11; // anything the ruleset couldn't match
 
     static const char* TYPE_MAP[];
